quase pronto

+++ fake_olx_springboot/src/main/java/com/ufms/olx/domain/dto/ProdutoDTO/ProdutoDTO.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -13,7 +15,7 @@ public class ProdutoDTO {
     private String descricao;
     private Long quantidadeEstoque;
     private Integer idadePermitida;
-    private double precoCompra;
-    private double precoVendaFisica;
-    private double precoVendaJuridica;
+    private BigDecimal precoCompra;
+    private BigDecimal precoVendaFisica;
+    private BigDecimal precoVendaJuridica;
 }
